Change the structure of the DrivetrainSubsystemBase class
I removed the Builder class and changed the masters and slaves and replaced them with basic implementation for each motor (master/slave) and create function that apply to all of them. by that, I simplified and modularized the subsystem, so it will be the same work, no matter how motors the drivetrain have.